Calculate performance usage on its own thread
We needed the information to be updated at a more frequent rate
than the one-second interval the main thread uses for refreshing
statistics.
