+++ onvm/onvm_mgr/main.c
@@ -69,6 +69,9 @@ static uint8_t worker_keep_running = 1;
 static void
 handle_signal(int sig);
 
+static void
+*rusage_thread_main(__attribute__((unused)) void *args);
+
 /*******************************Worker threads********************************/
 
 /*
@@ -100,15 +103,17 @@ master_thread_main(void) {
         /* Initial pause so above printf is seen */
         sleep(5);
 
+        pthread_t rusage_thd;
+        if (pthread_create(&rusage_thd, NULL, rusage_thread_main, NULL) != 0) {
+                RTE_LOG(WARNING, APP, "Unable to launch rusage thread.\n");
+        }
+
         onvm_stats_init(verbosity_level);
         /* Loop forever: sleep always returns 0 or <= param */
         while (main_keep_running && sleep(sleeptime) <= sleeptime) {
                 onvm_nf_check_status();
                 if (stats_destination != ONVM_STATS_NONE)
                         onvm_stats_display_all(sleeptime, verbosity_level);
-
-		// TODO Update stats here.
-		onvm_update_rusage();
 
                 if (time_to_live && unlikely((rte_get_tsc_cycles() - start_time) * TIME_TTL_MULTIPLIER /
                                              rte_get_timer_hz() >= time_to_live)) {
@@ -138,6 +143,9 @@ master_thread_main(void) {
 
         /* Stop all RX and TX threads */
         worker_keep_running = 0;
+
+        RTE_LOG(INFO, APP, "Core %d: Ending rusage thread\n", rte_lcore_id());
+        pthread_join(rusage_thd, NULL);
 
         /* Tell all NFs to stop */
         for (i = 0; i < MAX_NFS; i++) {
@@ -263,6 +271,14 @@ tx_thread_main(void *arg) {
         free(tx_mgr->tx_thread_info);
         free(tx_mgr->nf_rx_bufs);
         free(tx_mgr);
+        return 0;
+}
+
+static void
+*rusage_thread_main(__attribute__((unused)) void *args) {
+        while (worker_keep_running && usleep(RUSAGE_UPDATE_INTERVAL_MS * 1000) == 0) {
+                onvm_update_rusage();
+        }
         return 0;
 }
 
@@ -427,6 +443,7 @@ main(int argc, char *argv[]) {
                         }
                 }
         }
+
         /* Master thread handles statistics and NF management */
         master_thread_main();
         return 0;
